// src/app/reserves/model/reserves-table.model.ts
export class ReservesTable {
    reservesId: number;
    category: string;
    description: string;
    amount: number;
    currency: string;
    storageLocation: string;
    notice: String;
    createdDate: Date;
    attachment: boolean;
    attachmentPath: string;
    attachmentName: string;
    attachmentType: string;
}